+++ ec/utils.py
@@ -11,14 +11,14 @@ from modules.helpers import exit, gatherInput, reconfigArg
 
 __all__ = ['get', 'static', 'custom', 'walk', 'exit']
 
-def get(desc='', type=None, **ArgConfig):
+def get(type=None, **ArgConfig):
   r"""Helps to interactively get user input.
 
   Args:
     desc (str): The description for input.
     type (type / CustomType): The type of the input (defaults to None).
   """
-  ArgConfig.update(desc=desc, type=type)
+  ArgConfig.update(type=type)
   return gatherInput(**reconfigArg(ArgConfig))
 
 class custom(CustomType):